+++ projects/epgu-constructor/src/lib/component/unique-screen/components/confirm-personal-user-data-screen/confirm-personal-user-data.pipe.spec.ts
@@ -212,7 +212,7 @@ const resultOfRemoveFieldsOutsideJson = [
         value: '${birthPlace}',
       },
     ],
-    groupName: '${lastName} ${firstName} ${middleName}',
+    groupName: '${lastName} ${firstName} ${middleName} ${fullName}',
     visibilityLabel: '',
   },
   {
@@ -227,7 +227,12 @@ const resultOfRemoveFieldsOutsideJson = [
   },
   {
     fields: [],
-    groupName: 'Паспорт иностранного гражаина',
+    groupName: 'Паспорт иностранного гражданина',
+    visibilityLabel: '',
+  },
+  {
+    fields: [],
+    groupName: 'Свидетельство о рождении',
     visibilityLabel: '',
   },
 ];